Worked on the HomeScreen and ResultScreen
Added a background image to the HomeScreen.

Used a NavigationContainer to navigate between the
HomeScreen and ResultScreen.

Changed the button on the HomeScreen to navigate
to the ResultScreen.

Used touchablewithoutfeedback to hide the keyboard
when the user clicks outside of the input field.

Made it so the user inputs numbers only, and when
they put it into the input field, it will
that information is sent to the ResultScreen,
using useState.