Added new features
Added a fetch request to the weather API to get
the data from the server and display it on the
screen.

On the home screen, the user can enter the zip
code. The zip code is then used to fetch the
weather data from the server. The data is then
displayed on the result screen.

The result screen displays the weather data
fetched from the server. The user can go back
to the home screen by clicking the back button.