Updated logic
Whenever the apps fetchs a request and pulls
information based off of the weather, the
background changes on the result page to match the
weather type. For example, if the weather in 22406
is snowy, then the results page background will be
of a snow image.

Also, added images for the following weather types

- clear
- clouds
- fog
- rain
- snow