Added new images and new API link
Added smoke.png image to the weatherImages folder
and updated the API link in the HomeScreen
components.

Added a function for the weatherIcon in the
ResultScreen, so if the weather is clear, it
doesn't display an icon because it's not needed.
However, if the weather is anything other than
clear, it will display the appropriate icon.